refactor: deduplicate adapter parse functions into common module
Extract identical parse_reviewer_output/parse_reviewee_output and
summarize_json/summarize_text from claude.rs and codex.rs into
common.rs. Also eliminate continue_session duplication in claude.rs
by unifying it with run_claude_streaming via Optional allowed_tools.
